+++ crates/teloxide-core/CHANGELOG.md
@@ -71,7 +71,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Add `ExpandableBlockquote` message entity
 
 - Support for TBA 7.5 ([#1281](pr1281))
-  - Add `TransactionPartner` and `RevenueWithdrawalState` structs
+  - Add `StarTransactions`, `StarTransaction`, `TransactionPartner` and `RevenueWithdrawalState` structs
 
 [pr1157]: https://github.com/teloxide/teloxide/pull/1157
 [pr1264]: https://github.com/teloxide/teloxide/pull/1264